academic_twitch_v2/requirements.txt
twitchio>=3.1.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
cachetools>=5.3.0
rapidfuzz>=3.6.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
discord.py>=2.4.0
